+++ mayan/apps/cabinets/icons.py
@@ -4,5 +4,6 @@ from appearance.classes import Icon
 
 icon_cabinet = Icon(driver_name='fontawesome', symbol='columns')
 icon_cabinet_add = Icon(driver_name='fontawesome', symbol='plus')
+icon_cabinet_child_add = Icon(driver_name='fontawesome', symbol='plus')
 icon_cabinet_create = Icon(driver_name='fontawesome', symbol='plus')
 icon_cabinet_list = Icon(driver_name='fontawesome', symbol='columns')